move layout to Layout (for naming consistency)
